ml/quality/BRISQUE+OpenCLIP/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
piq>=0.8.0
open-clip-torch>=2.20.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data manipulation  
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# HTTP requests
requests>=2.31.0

# Machine learning (if fine-tuning attempted)
scikit-learn>=1.3.0
timm>=0.9.0

# Optional: GPU acceleration
# nvidia-cuda-runtime-cu11>=11.7  # Uncomment if CUDA available
